cleanup: the small half of the review, including a key that meant two things
All 29 goldens 0 px, transform arms green.

The probe's total row said moved= where the node rows say moved=, and the two
were different quantities -- a boolean per node against a count for the frame.
A reader filtering on it alone would have picked up the totals and found no step
key, which is the hazard _probe_rows' own comment warns about. It is movers= now.

That row also gains nodes=, because the probe only emits a row for a NAMED node
and a scene loaded from a GLB is mostly unnamed. Without the full count the arms
were reasoning about a subset with no way to tell.

The walk carried a defensive copy of the parent matrix on every pop. The only
read of it happens before any child push can overwrite the slot, and the
multiply's destination is on the node rather than the stack, so it cannot alias
its own source. 64 bytes per node per frame, about 580 KB on forest, that
nothing read.

shift_node_tree's prev half is dead now rather than wrong: the latch runs between
the shift and the walk and overwrites it. Kept, because it is what makes that
function correct on its own instead of by where the engine calls it, and said so.

Also: glm_vec3_distance rather than a subtract into a temporary and a norm; the
probe's frame widened to the size_t it is actually fed, so the call site stops
casting; _probe_render's docstring paragraphs moved onto _probe_rows, which is
the function they have described since the split; and the blank lines the
callback split left before three closing braces.

+++ apps/spores/src/spores.c
@@ -314,7 +314,7 @@ static void on_render(Game* game, double alpha) {
     // The engine walked the graph before the shadow pass, so this reads what
     // that walk produced -- which is the only point the answer means anything.
     if (g_transform_probe > 0 && (int)engine->total_frames % g_transform_probe == 0)
-        scene_transform_probe(scene, (int)engine->total_frames);
+        scene_transform_probe(scene, engine->total_frames);
 
     // The scene's particle systems are ticked + drawn by the engine; nothing to
     // do here but render the scene.
